requirements.txt
PyYAML>=6.0.2
psycopg[binary]>=3.2.1
neo4j>=5.28.1
matplotlib>=3.9.2
networkx>=3.3
numpy>=2.1.1
pulp>=2.9.0
